refactor: moved `RelativeCoord` into render_implementations/coords.rs
